typecast the return
git-svn-id: svn://localhost/svn/uclinux-dist/trunk@3988 3e2a5524-39e8-4cb9-84a9-c49da69cba3a

+++ linux-2.6.x/drivers/mtd/chips/map_ram.c
@@ -97,7 +97,7 @@ static unsigned long mapram_unmapped_area(struct mtd_info *mtd,
 					  unsigned long flags)
 {
 	struct map_info *map = mtd->priv;
-	return map->virt + offset;
+	return (unsigned long)(map->virt + offset);
 }
 
 static int mapram_read (struct mtd_info *mtd, loff_t from, size_t len, size_t *retlen, u_char *buf)
